requirements_minimal.txt
python-telegram-bot==20.7
groq>=0.9.0
python-dotenv==1.0.0
requests==2.31.0

